+++ app/Http/Controllers/TaskController.php
@@ -7,6 +7,8 @@ use App\Http\Requests\TaskRequest;
 use App\Models\Client;
 use App\Models\Project;
 use App\Models\User;
+use App\PermissionEnum;
+use Illuminate\Support\Facades\Gate;
 
 class TaskController extends Controller
 {
@@ -41,6 +43,7 @@ class TaskController extends Controller
     }
     public function destroy(Task $task)
     {
+        Gate::authorize(PermissionEnum::DELETE_TASK->value);
         $task->delete();
         return redirect()->route('tasks.index')->with('success', 'Task deleted successfully.');
     }
